Fix Gemfile.lock for Depfu
Thanks for using Depfu on this repo. Unfortunately one of the latest commits to this file added the BUNDLED WITH block with a missing line break at the end which breaks Depfu's version parser to decide which bundler version to run against the project. You could argue that we should lessen the strictness of our parser but Gemfile.lock is usually autogenerated and therefore we never run into this problem unless someone messes with it by hand.

There's two things of note here: 
1. Please don't edit Gemfile.lock files by hand. They are always auto generated by Bundler and messing with them in an editor usually breaks things in very subtle ways
2. Textfiles should always end on a newline :)